Organize fields into files. Bump version number. Bump dependency versions.
