+++ lib/models/field_types/championtextfield.dart
@@ -1,0 +1,77 @@
+import 'package:championforms/models/field_types/formfieldclass.dart';
+import 'package:flutter/material.dart';
+
+class ChampionTextField extends FormFieldDef {
+  // Define the type of field type
+
+  // Add a TextField override so we could write our own widget if we prefer. This will override the default field.
+  final TextField? fieldOverride;
+
+  final int? maxLines;
+
+  // Add a title to the text field itself if desired
+  final String? textFieldTitle;
+
+  // Add hint text if needed
+  final String hintText;
+
+  final Widget? leading;
+  final Widget? trailing;
+
+  // Does this field have a max length?
+  final int? maxLength;
+
+  // obfuscate the field
+  final bool password;
+
+  // These are the default values for the field. Use the specific one you need depending on the input required.
+  final String? defaultValue;
+
+  // Add a builder for defining the field style
+
+  // We need to have a callback which will be called when drag and drop
+  final Future<void> Function({
+    TextEditingController controller,
+    required String formId,
+    required String fieldId,
+  })? onDrop;
+
+  // Does this field support drag functionality?
+  final bool draggable;
+
+  // We need to have a callback which will be called when content is pasted
+  final Future<void> Function({
+    TextEditingController controller,
+    required String formId,
+    required String fieldId,
+  })? onPaste;
+
+  ChampionTextField({
+    required super.id,
+    this.fieldOverride,
+    this.maxLines,
+    this.textFieldTitle,
+    this.hintText = "",
+    super.icon,
+    this.leading,
+    this.trailing,
+    super.theme,
+    super.title,
+    super.description,
+    this.maxLength,
+    super.disabled,
+    super.hideField,
+    super.requestFocus,
+    this.password = false,
+    this.defaultValue,
+    super.validators,
+    super.validateLive,
+    super.onSubmit,
+    super.onChange,
+    this.onDrop,
+    this.draggable = true,
+    this.onPaste,
+    super.fieldLayout,
+    super.fieldBackground,
+  });
+}
