+++ lib/models/field_types/formfieldclass.dart
@@ -1,5 +1,6 @@
 import 'package:championforms/controllers/form_controller.dart';
 import 'package:championforms/models/colorscheme.dart';
+import 'package:championforms/models/field_types/championoptionselect.dart';
 import 'package:championforms/models/fieldstate.dart';
 import 'package:championforms/models/formbuildererrorclass.dart';
 import 'package:championforms/models/field_types/formfieldbase.dart';
@@ -102,175 +103,5 @@ abstract class FormFieldDef implements FormFieldBase {
     this.fieldLayout = fieldSimpleLayout, // Default to the simple layout
     this.fieldBackground =
         fieldSimpleBackground, // Default to the simple (no) field background
-  });
-}
-
-class ChampionTextField extends FormFieldDef {
-  // Define the type of field type
-
-  // Add a TextField override so we could write our own widget if we prefer. This will override the default field.
-  final TextField? fieldOverride;
-
-  final int? maxLines;
-
-  // Add a title to the text field itself if desired
-  final String? textFieldTitle;
-
-  // Add hint text if needed
-  final String hintText;
-
-  final Widget? leading;
-  final Widget? trailing;
-
-  // Does this field have a max length?
-  final int? maxLength;
-
-  // obfuscate the field
-  final bool password;
-
-  // These are the default values for the field. Use the specific one you need depending on the input required.
-  final String? defaultValue;
-
-  // Add a builder for defining the field style
-
-  // We need to have a callback which will be called when drag and drop
-  final Future<void> Function({
-    TextEditingController controller,
-    required String formId,
-    required String fieldId,
-  })? onDrop;
-
-  // Does this field support drag functionality?
-  final bool draggable;
-
-  // We need to have a callback which will be called when content is pasted
-  final Future<void> Function({
-    TextEditingController controller,
-    required String formId,
-    required String fieldId,
-  })? onPaste;
-
-  ChampionTextField({
-    required super.id,
-    this.fieldOverride,
-    this.maxLines,
-    this.textFieldTitle,
-    this.hintText = "",
-    super.icon,
-    this.leading,
-    this.trailing,
-    super.theme,
-    super.title,
-    super.description,
-    this.maxLength,
-    super.disabled,
-    super.hideField,
-    super.requestFocus,
-    this.password = false,
-    this.defaultValue,
-    super.validators,
-    super.validateLive,
-    super.onSubmit,
-    super.onChange,
-    this.onDrop,
-    this.draggable = true,
-    this.onPaste,
-    super.fieldLayout,
-    super.fieldBackground,
-  });
-}
-
-class ChampionOptionSelect extends FormFieldDef {
-  // Define the type of field type
-
-  final Widget? leading;
-  final Widget? trailing;
-
-  // Options
-  final List<MultiselectOption> options;
-
-  // Multiselect?
-  final bool multiselect;
-
-  // These are the default values for the field. Use the specific one you need depending on the input required.
-  final List<String> defaultValue;
-
-  // match default value case sensitive?
-  final bool caseSensitiveDefaultValue;
-
-  // Add a builder for defining the field style
-  Widget Function(
-    BuildContext context,
-    ChampionFormController controller,
-    List<MultiselectOption> choices,
-    ChampionOptionSelect field,
-    FieldState currentState,
-    FieldColorScheme currentColors,
-    List<String>? defaultValue,
-    Function(bool focused) updateFocus,
-    Function(MultiselectOption? selectedOption) updateSelectedOption,
-  ) fieldBuilder;
-
-  ChampionOptionSelect({
-    required super.id,
-    super.icon,
-    required this.options,
-    this.multiselect = false,
-    this.leading,
-    this.trailing,
-    super.theme,
-    super.title,
-    super.description,
-    super.disabled,
-    super.hideField,
-    super.requestFocus,
-    this.defaultValue = const [],
-    this.caseSensitiveDefaultValue = true,
-    super.validators,
-    super.validateLive,
-    super.onSubmit,
-    super.onChange,
-    super.fieldLayout,
-    super.fieldBackground,
-    this.fieldBuilder = dropdownFieldBuilder,
-  });
-}
-
-class ChampionCheckboxSelect extends ChampionOptionSelect {
-  // Add a builder for defining the field style
-  @override
-  Widget Function(
-    BuildContext context,
-    ChampionFormController controller,
-    List<MultiselectOption> choices,
-    ChampionOptionSelect field,
-    FieldState currentState,
-    FieldColorScheme currentColors,
-    List<String>? defaultValue,
-    Function(bool focused) updateFocus,
-    Function(MultiselectOption? selectedOption) updateSelectedOption,
-  ) fieldBuilder;
-  ChampionCheckboxSelect({
-    required super.id,
-    super.icon,
-    required super.options,
-    super.multiselect = false,
-    super.leading,
-    super.trailing,
-    super.theme,
-    super.title,
-    super.description,
-    super.disabled,
-    super.hideField,
-    super.requestFocus,
-    super.defaultValue = const [],
-    super.caseSensitiveDefaultValue = true,
-    super.validators,
-    super.validateLive,
-    super.onSubmit,
-    super.onChange,
-    super.fieldLayout,
-    super.fieldBackground,
-    this.fieldBuilder = checkboxFieldBuilder,
   });
 }
