+++ lib/widgets_internal/field_widgets/file_upload_widget.dart
@@ -2,9 +2,9 @@ import 'package:championforms/controllers/form_controller.dart';
 import 'package:championforms/functions/filetype_from_mime.dart';
 import 'package:championforms/models/field_types/championfileupload.dart';
 import 'package:championforms/models/colorscheme.dart';
+import 'package:championforms/models/field_types/championoptionselect.dart';
 import 'package:championforms/models/fieldstate.dart';
 import 'package:championforms/models/file_model.dart';
-import 'package:championforms/models/field_types/formfieldclass.dart';
 import 'package:championforms/models/formresults.dart';
 import 'package:championforms/models/mime_filetypes.dart';
 import 'package:championforms/models/multiselect_option.dart';
