+++ lib/widgets_internal/field_widgets/textfieldwidget.dart
@@ -1,6 +1,7 @@
 import 'package:championforms/controllers/form_controller.dart';
 import 'package:championforms/fieldbuilders/textfieldbuilder.dart';
 import 'package:championforms/models/colorscheme.dart';
+import 'package:championforms/models/field_types/championtextfield.dart';
 import 'package:championforms/models/fieldstate.dart';
 import 'package:championforms/models/field_types/formfieldclass.dart';
 import 'package:championforms/models/formresults.dart';
