+++ lib/widgets_internal/formbuilder.dart
@@ -2,7 +2,9 @@ import 'package:championforms/controllers/form_controller.dart';
 import 'package:championforms/functions/gather_child_errors.dart';
 import 'package:championforms/models/colorscheme.dart';
 import 'package:championforms/models/field_types/championcolumn.dart';
+import 'package:championforms/models/field_types/championoptionselect.dart';
 import 'package:championforms/models/field_types/championrow.dart';
+import 'package:championforms/models/field_types/championtextfield.dart';
 import 'package:championforms/models/fieldstate.dart';
 import 'package:championforms/models/formresults.dart';
 import 'package:championforms/models/multiselect_option.dart';
