+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: championforms
 description: "Easy Build Data Input Fields of All Types in Flutter"
-version: 0.0.4
+version: 0.0.5
 homepage: "https://championsoftheweb.com"
 repository: "https://github.com/cotw-fabier/championforms/"
 
@@ -12,13 +12,11 @@ dependencies:
   flutter:
     sdk: flutter
   email_validator: ^3.0.0
-  super_clipboard: ^0.8.17
-  super_drag_and_drop: ^0.8.16
-  dotted_border: ^2.1.0
-  textfield_tags: ^3.0.1
+  super_clipboard: ^0.8.24
+  super_drag_and_drop: ^0.8.24
   collection: ^1.19.0
   uuid: ^4.5.1
-  file_picker: ^8.1.7
+  file_picker: ^8.3.5
   cross_file: ^0.3.4+2
   mime: ^2.0.0
